+++ lib/rubocop/cop/sensei_cops.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'lecture/explain_bang_bang'
+require_relative 'lecture/explain_lonely_operator'
 require_relative 'lecture/explain_numbered_parameters'
 require_relative 'lecture/explain_or_assign'
 require_relative 'lecture/explain_pattern_match_case_in'